Minor bugfix
There was a bug that involves Dungeon, Miniboss, and Boss entrance shuffling where after an entrance was marked, the list of location checks would not show up at all unless you entered the dungeon, miniboss, or boss room for the second time. This felt weird to me and in order to avoid this situation from happening with other people, I just simply moved the code where the checks for each location show up on the sector portion of the tracker to the very bottom of a conditional where if a user choose to randomize dungeon entrances, code in it would be executed unless comments were included in which case that part of the code would be ignored.